Add null check to DictionaryTransformer
If some type GUID is unknown to Attic, default(T) will be returned for
the key. For reference types and nullable value types, the key will be
null and Add will throw an ArgumentNullException. This commit prevents
that. However, unknown value types are not considered yet, and Add will
throw an ArgumentException once the transformer tries to add another
key-value pair with a key that has already been added.